pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scrapbook-md"
version = "1.0.0"
description = "A lightweight CLI tool for capturing ideas, prompts, and todos in markdown format"
authors = [{name = "Developer", email = "dev@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "click>=8.0.0",
    "PyYAML>=6.0.0"
]

[project.scripts]
scrap = "cli.cli:main"

[project.urls]
Homepage = "https://github.com/user/scrapbook-md"
Repository = "https://github.com/user/scrapbook-md"

[tool.setuptools.packages.find]
where = ["."]
include = ["cli*"]